+++ scripts/force-scilla-download.ts
@@ -1,3 +1,3 @@
-import { parseScilla } from "../src/ScillaParser"
+import { parseScilla } from "../src/parser/ScillaParser"
 
 parseScilla("test/fixture-projects/hardhat-project/contracts/scilla/HelloWorld.scilla");
