+++ src/parser/ScillaContractsInfoUpdater.ts
